src/utils/supabase_client.py
from __future__ import annotations

import os
from functools import lru_cache
from typing import Optional

from dotenv import load_dotenv
from supabase import Client, create_client


load_dotenv()


@lru_cache()
def get_supabase_client() -> Client:
    url = os.getenv("SUPABASE_URL")
    key = os.getenv("SUPABASE_KEY")
    if not url or not key:
        raise RuntimeError("Supabase credentials are not configured in the environment")
    return create_client(url, key)


__all__ = ["get_supabase_client"]
