requirements.txt
python-dotenv>=1.0
pydantic>=2.8
PyYAML>=6.0
httpx>=0.27
tqdm>=4.66
loguru>=0.7
supabase>=2.6
boto3>=1.34
openai>=1.51.0
elevenlabs>=1.6.1
mutagen>=1.47.0
pydub>=0.25.1
